requirements.txt
black
pylint
radon
boto3
litellm
python-dotenv